+++ package.json
@@ -69,6 +69,7 @@
   "homepage": "https://github.com/dy/spect#readme",
   "type": "module",
   "dependencies": {
+    "delegate-it": "^1.1.0",
     "delegated-events": "^1.1.1",
     "fast-deep-equal": "^2.0.1",
     "fast-on-load": "^1.2.0",
@@ -82,6 +83,7 @@
     "negative-index": "^1.0.3",
     "preact": "^10.0.0-rc.3",
     "primitive-pool": "^1.1.0",
+    "regular-elements": "^0.5.0",
     "scope-css": "^1.2.1",
     "selector-observer": "^2.1.6"
   },
