+++ src/index.js
@@ -1,6 +1,8 @@
 import use from './use'
 import on, { fire } from './on'
 import { run } from './core'
+import fx from './fx'
+import $, { $$ } from './$'
 import attr from './attr'
 import state from './state'
 import prop from './prop'
@@ -9,4 +11,4 @@ import data from './data'
 import query from './query'
 
 
-export { use, on, fire, state, run }
+export { use, on, fire, state, run, fx, prop, $, $$ }
