+++ src/prop.js
@@ -1,5 +1,5 @@
 import {createEffect} from './util'
 
 export default createEffect(function prop(target) {
-  return el
+  return target
 })
