+++ src/state.js
@@ -2,8 +2,8 @@ import tuple from 'immutable-tuple'
 import { createEffect } from './util'
 
 const stateCache = new WeakMap
-export default createEffect(function prop(el) {
-  let state = stateCache.get(tuple(el))
-  if (!state) stateCache.set(tuple(el), state = {})
+export default createEffect(function state(target) {
+  let state = stateCache.get(tuple(target))
+  if (!state) stateCache.set(tuple(target), state = {})
   return state
 })
