+++ test/use.js
@@ -1,7 +1,32 @@
 import t from 'tst'
-import $ from '..'
+import { use } from '..'
 
-t('use: aspects, assigned through parent wrapper', async t => {
+t('use: observe selector', async t => {
+  let log = []
+
+  use('.item', el => {
+    log.push(el)
+    return () => {
+      log.push(el)
+    }
+  })
+
+  let el = document.createElement('a')
+  el.classList.add('item')
+  document.body.appendChild(el)
+
+  await ''
+
+  t.is(log, [el])
+
+  document.body.removeChild(el)
+
+  await ''
+
+  t.is(log, [el, el])
+})
+
+t.todo('use: aspects, assigned through parent wrapper', async t => {
   // some wrappers over same elements can be created in different ways
   // but effects must be bound to final elements
   let a = document.createElement('a')
@@ -28,7 +53,7 @@ t('use: aspects, assigned through parent wrapper', async t => {
   t.is($a5[0], a)
 })
 
-t('use: aspects must be called in order', async t => {
+t.todo('use: aspects must be called in order', async t => {
   let log = []
   let a = document.createElement('a')
   await $(a).use([() => log.push(1), () => log.push(2), () => log.push(3)])
@@ -36,7 +61,7 @@ t('use: aspects must be called in order', async t => {
   t.deepEqual(log, [1,2,3])
 })
 
-t('use: duplicates are ignored', async t => {
+t.todo('use: duplicates are ignored', async t => {
   let log = []
 
   await $(document.createElement('a')).use([fn, fn, fn])
@@ -52,7 +77,7 @@ t('use: duplicates are ignored', async t => {
   t.is(log, [1, 1])
 })
 
-t('use: aspects must not be called multiple times, unless target state changes', async t => {
+t.todo('use: aspects must not be called multiple times, unless target state changes', async t => {
   let log = []
 
   let $a = $`<a/>`
@@ -68,7 +93,7 @@ t('use: aspects must not be called multiple times, unless target state changes',
   function fn(el) {log.push('x')}
 })
 
-t('use: same aspect different targets', async t => {
+t.todo('use: same aspect different targets', async t => {
   let log = []
   function fx(el) {
     log.push(el.tagName)
@@ -88,7 +113,7 @@ t('use: same aspect different targets', async t => {
   t.deepEqual(log, ['A', 'SPAN'])
 })
 
-t('use: Same target different aspects', async t => {
+t.todo('use: Same target different aspects', async t => {
   let log = []
 
   let a = document.createElement('a')
@@ -103,7 +128,7 @@ t('use: Same target different aspects', async t => {
   document.body.removeChild(a)
 })
 
-t('use: same aspect same target', async t => {
+t.todo('use: same aspect same target', async t => {
   let log = []
   let a = document.createElement('a')
   document.body.appendChild(a)
@@ -119,7 +144,7 @@ t('use: same aspect same target', async t => {
   document.body.removeChild(a)
 })
 
-t('use: subaspects init themselves independent of parent aspects', async t => {
+t.todo('use: subaspects init themselves independent of parent aspects', async t => {
   let log = []
 
   let a = document.body.appendChild(document.createElement('a'))
@@ -169,3 +194,146 @@ t.todo('use: new custom element', t => {
   })
 })
 
+
+
+t.todo('use: aspects must be called in order', async t => {
+  let log = []
+  let a = {}
+  await spect(a).use([() => log.push(1), () => log.push(2), () => log.push(3)])
+  t.deepEqual(log, [1, 2, 3])
+})
+
+t.todo('use: duplicates are ignored', async t => {
+  let log = []
+
+  await spect({}).use([fn, fn, fn])
+
+  function fn() {
+    log.push(1)
+  }
+
+  t.is(log, [1])
+
+  await spect({}).use([fn, fn, fn])
+
+  t.is(log, [1, 1])
+})
+
+t.todo('use: aspects must not be called multiple times, unless target state changes', async t => {
+  let log = []
+
+  let $a = spect({})
+  await $a.use(fn)
+  t.is(log, ['x'])
+  await $a.use(fn)
+  t.is(log, ['x'])
+  await $a.use([fn, fn])
+  t.is(log, ['x'])
+  await $a.update()
+  t.is(log, ['x', 'x'])
+
+  function fn(el) { log.push('x') }
+})
+
+t.skip('use: same aspect different targets', t => {
+  let log = []
+  function fx([el]) {
+    log.push(el.tagName)
+    // return () => log.push('destroy ' + el.tagName)
+  }
+
+  let $el = spect({ tagName: 'A' }).use(fx)
+
+  t.is($el.target.tagName, log[0])
+  t.is(log, ['A'])
+
+  $el.target.innerHTML = '<span></span>'
+  $($el.target.firstChild).use(fx)
+
+  t.deepEqual(log, ['A', 'SPAN'])
+})
+
+t.todo('use: Same target different aspects', async t => {
+  let log = []
+
+  let a = {}
+
+  let afx, bfx
+  await spect(a).use([afx = () => (log.push('a'), () => log.push('de a'))])
+  t.deepEqual(log, ['a'])
+  await spect(a).use([bfx = () => (log.push('b'), () => log.push('de b'))])
+  t.deepEqual(log, ['a', 'b'])
+})
+
+t.todo('use: same aspect same target', async t => {
+  let log = []
+  let a = {}
+
+  let fx = () => (log.push('a'), () => log.push('z'))
+  await spect(a).use(fx)
+  t.deepEqual(log, ['a'])
+  await spect(a).use(fx)
+  t.deepEqual(log, ['a'])
+  await spect(a).use(fx)
+  t.deepEqual(log, ['a'])
+})
+
+t.todo('use: subaspects init themselves independent of parent aspects', async t => {
+  let log = []
+
+  let a = { b: { c: {} } }
+  let b = a.b
+  let c = b.c
+
+  await spect(a).use(el => {
+    log.push('a')
+    spect(b).use(el => {
+      log.push('b')
+      spect(c).use(el => {
+        log.push('c')
+        // return () => log.push('-c')
+      })
+      // return () => log.push('-b')
+    })
+    // return () => log.push('-a')
+  })
+
+  t.deepEqual(log, ['a', 'b', 'c'])
+
+  // $.destroy(a)
+
+  // t.deepEqual(log, ['a', 'b', 'c', '-c', '-b', '-a'])
+})
+
+t.todo('use: generators aspects')
+
+t.todo('use: async aspects', t => {
+  let a = spect({})
+
+  a.use(async function a() {
+    t.is(a, current.fn)
+    await Promise.resolve().then()
+    t.is(a, current.fn)
+  })
+
+})
+
+t.skip('use: promise', async t => {
+  let to = new Promise(ok => setTimeout(ok, 100))
+
+  to.then()
+
+  spect({}).use(to)
+})
+
+
+t.todo('fx: global effect is triggered after current callstack', async t => {
+  let log = []
+  spect({}).fx(() => log.push('a'))
+
+  t.is(log, [])
+
+  await 0
+
+  t.is(log, ['a'])
+})
